Add cached token provider

+++ modules/stackdriver-http-exporter/src/main/scala/io/janstenpickle/trace4cats/strackdriver/oauth/AccessToken.scala
@@ -6,7 +6,7 @@ Code adapted from https://github.com/permutive/fs2-google-pubsub
 import io.circe.Codec
 import io.circe.generic.extras.semiauto._
 
-final case class AccessToken(accessToken: String, tokenType: String, expiresIn: Int)
+final case class AccessToken(accessToken: String, tokenType: String, expiresIn: Long)
 
 object AccessToken {
   implicit val codec: Codec[AccessToken] = deriveConfiguredCodec
